Dockerfile: Add BUILD_ID build argument
This can be used to break the cache when we know the upstream package has
changed. In CI we will probably just put the build number here.

Even though it is not currently exposed anywhere, it is available as an
environment variable in all RUN commands following the ARG. This should be
enough to avoid reusing cached layers between builds.